% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarymcmc}
\alias{summarymcmc}
\title{summary of output from `atmcmc'}
\usage{
summarymcmc(output, name = "MCMC")
}
\arguments{
\item{output}{output from an `atmcmc' run}

\item{name}{name of the project}
}
\value{
Displays the following items:
 \item{name}{name of the project}
 \item{estimates}{sample mean}
 \item{estimates_of_functional}{functional sample mean}
 \item{acceptance_rate}{acceptance rate of the 2nd half of sampling phase}
 \item{time_elapsed}{runtime of the full MCMC}
 \item{phase_length}{string of iteration numbers to show when each phase has ended. For the sampling phase, this shows when the 1st half of sampling phase has ended also}
}
\description{
Shows a summary of output from an `atmcmc' run. It includes project name, sample mean, functional sample mean, acceptance rate, runtime, and iteration numbers of end of all phases.
}
\examples{
## see examples in `atmcmc'
}

